fix: switch to grid layout
