+++ src/component/Sidebar/SidebarContent.css
@@ -1,31 +1,14 @@
 aside.sidebar {
-  position: relative;
-  margin-left: calc(var(--base-size-navbar) - 1px);
-  width: auto;
-  z-index: 40;
   background-color: var(--bg-blue-3);
-  padding-bottom: calc(var(--gap) / 2);
-  max-height: calc(100vh - var(--base-size-navbar));
 }
 
 @media (min-width: 400px) {
-  aside.sidebar {
-    position: fixed;
-    left: calc(var(--base-size-navbar) - 1px);
-    top: var(--base-size-navbar);
-    max-height: 100vh;
-    width: calc(var(--size-sidebar) - var(--base-size-navbar));
-    border-top-right-radius: calc(var(--base-size-navbar) / 4);
-    border-bottom-right-radius: calc(var(--base-size-navbar) / 4);
-    margin-left: 0;
-  }
-
   aside.sidebar > div.scrollable {
     overflow-y: auto;
     overflow-x: hidden;
-    max-height: calc(100vh - var(--base-size-navbar) * 2);
     scrollbar-color: var(--bg-blue-5) var(--bg-blue-1);
     scrollbar-width: thin;
+    max-height: calc(100% - var(--base-size-navbar));
   }
 }
 
@@ -40,9 +23,11 @@ aside.sidebar > header {
 aside.sidebar > div.scrollable > section {
   margin: 0 var(--gap);
 }
+
 aside.sidebar > div.scrollable > section + section {
   margin-top: var(--gap);
 }
+
 aside.sidebar > div.scrollable > footer {
   border-top-width: 1px;
   border-top-style: solid;
@@ -52,6 +37,7 @@ aside.sidebar > div.scrollable > footer {
   opacity: 0.8;
   font-size: 90%;
 }
+
 aside.sidebar > div.scrollable > section.separator {
   border-top-width: 1px;
   border-top-style: solid;
