+++ src/component/SidebarNav.tsx
@@ -1,7 +1,8 @@
 import { SidebarButton as DeviceDetailButton } from './Device.js'
 import { SidebarButton as ViewSourceButton } from './ViewSource.js'
 import { SidebarButton as AddDeviceButton } from './AddDevice.js'
-import { WarningButton, SearchButton } from './Sidebar.js'
+import { SearchButton } from './Sidebar/SearchButton.tsx'
+import { WarningButton } from './Sidebar/WarningButton.tsx'
 import { SidebarButton as TutorialButton } from './Tutorial/Tutorial.tsx'
 
 import './SidebarNav.css'
