+++ src/component/Tutorial/isDone.ts
@@ -5,7 +5,7 @@ import { encode } from '../../context/navigation/encodeNavigation.ts'
 export const isDone = (
 	tutorial: TutorialEntryType,
 	location: ReturnType<typeof useNavigation>,
-) => {
+): boolean => {
 	const done = tutorial.done
 	if (done === undefined) return false
 	if ((encode(location.current())?.includes(done) ?? false) === false)
