+++ src/component/WIPSidebar.tsx
@@ -1,6 +1,6 @@
 import { useNavigation } from '../context/Navigation.js'
 import { Close } from '../icons/LucideIcon.js'
-import { SidebarContent } from './Sidebar.js'
+import { SidebarContent } from './Sidebar/SidebarContent.tsx'
 import { Show } from 'solid-js'
 
 export const WIPSidebar = () => {
